+++ tests/packageData.test.js
@@ -1,4 +1,5 @@
 import packageData, { CACHE } from '../src/packageData';
+import { SEQUENCE_REMOVAL_PATTERN } from '../src/constants';
 
 describe('packageData()', () => {
   it('caches stable data', () => {
@@ -15,7 +16,7 @@ describe('packageData()', () => {
 
   packageData().forEach((emoji) => {
     it(`formats a hexcode without ZWJ for ${emoji.unicode}`, () => {
-      expect(emoji.hexcode).not.toMatch(/(200D|FE0F)/);
+      expect(emoji.hexcode).not.toMatch(SEQUENCE_REMOVAL_PATTERN);
     });
 
     it(`hexcode doesnt exceed length 4 for ${emoji.unicode}`, () => {
@@ -47,7 +48,10 @@ describe('packageData()', () => {
     });
 
     it(`defines an order for ${emoji.unicode}`, () => {
-      expect(emoji.order).toBeGreaterThanOrEqual(1);
+      // Newer emoji may not have an order yet
+      if (emoji.order) {
+        expect(emoji.order).toBeGreaterThanOrEqual(1);
+      }
     });
 
     it(`defines tags for ${emoji.unicode}`, () => {